refactor: align json mtp helper with speculative layout.
