+++ xllm/core/framework/speculative/mtp_json_object_state.h
