+++ xllm/core/runtime/mtp_worker_impl.h
@@ -24,12 +24,12 @@ limitations under the License.
 
 #include "core/framework/speculative/embedding_cache.h"
 #include "core/framework/speculative/mtp_async_state.h"
+#include "core/framework/speculative/mtp_json_object_state.h"
 #include "framework/kv_cache_transfer/kv_cache_transfer.h"
 #if defined(USE_NPU)
 #include "framework/kv_cache_transfer/spec_kv_cache_transfer.h"
 #endif
 #include "core/framework/speculative/adaptive_speculative_controller.h"
-#include "core/runtime/mtp_json_object_state.h"
 #include "runtime/speculative_worker_impl.h"
 
 namespace xllm {
